+++ src/control_plane/migration/mod.rs
@@ -2,6 +2,7 @@
 mod tests;
 
 mod migration_backup;
+mod migration_cutover_plan;
 mod migration_error;
 mod migration_execution_result;
 mod migration_operation_error;
@@ -10,6 +11,7 @@ mod postgres;
 mod run_migration;
 
 pub(crate) use migration_backup::MigrationBackup;
+pub(crate) use migration_cutover_plan::MigrationCutoverPlan;
 pub(crate) use migration_error::MigrationError;
 pub(crate) use migration_execution_result::MigrationExecutionResult;
 pub(crate) use migration_operation_error::MigrationOperationError;